machine-learning/datasets/dla/pyproject.toml
[project]
name = "dla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.7",
    "chunkr-ai>=0.0.48",
    "nomic>=3.5.0",
    "psycopg[binary]>=3.2.7",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.45.0",
]
